fix rlp encoding issues
