+++ lib/rly_network_flutter_sdk.dart
@@ -4,3 +4,5 @@ export 'src/wallet_manager.dart';
 export 'src/wallet.dart';
 export 'src/gsn/meta_tx_method.dart';
 export 'src/token_helpers.dart';
+export 'src/zksync/eip712Transaction.dart';
+export 'src/zksync/sendEip712Transaction.dart';
