+++ lib/src/gsn/utils.dart
@@ -1,5 +1,7 @@
 import 'package:http/http.dart';
 import 'package:web3dart/web3dart.dart';
+import 'dart:convert';
+import 'dart:typed_data';
 
 enum RlyEnv {
   local,
@@ -60,4 +62,30 @@ typedef IntString = String;
 Web3Client getEthClient(String apiUrl) {
   var httpClient = Client();
   return Web3Client(apiUrl, httpClient);
+}
+
+String concatHex(List<String> values) {
+    String concatenatedHex = values.fold('', (acc, x) => acc + x.replaceAll('0x', ''));
+    String hexWithPrefix = '0x' + concatenatedHex;
+  return hexWithPrefix;
+}  
+
+Uint8List hexToUint8List(String hex) {
+  // Ensure the hex string has an even length
+  if (hex.length % 2 != 0) {
+    throw FormatException("Invalid hex string");
+  }
+
+  // Remove the "0x" prefix if present
+  if (hex.startsWith("0x")) {
+    hex = hex.substring(2);
+  }
+
+  // Convert the hex string to a Uint8List
+  final List<int> bytes = [];
+  for (int i = 0; i < hex.length; i += 2) {
+    bytes.add(int.parse(hex.substring(i, i + 2), radix: 16));
+  }
+
+  return Uint8List.fromList(bytes);
 }
